Fix task 1

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,4 +1,5 @@
 #include "function_pointers.h"
+#include <stddef.h>
 /**
 * array_iterator - executes a function given as a parameter on each element of an array
 * @array: integer array
@@ -9,7 +10,7 @@
 */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-int index;
+long unsigned int index;
 
 for (index = 0; index < size; index++)
 {
